refactor: Defender 클래스를 Defender, ScoreKeeper 클래스로 분리 및 참조하는 클래스 수정
