+++ src/main/java/baseball/Session.java
@@ -10,7 +10,7 @@ public class Session {
             String answer = Console.readLine();
 
             Validation.isValidAnswerTarget(answer);
-            Defender.printScore(answer);
+            ScoreKeeper.printScore(answer);
 
         } while (!endGame());
     }
@@ -20,7 +20,7 @@ public class Session {
         boolean isRestart = false;
         String endMessage = "3개의 숫자를 모두 맞히셨습니다! 게임 종료\n게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.";
 
-        if (Defender.isThreeStrikes()) {
+        if (ScoreKeeper.isThreeStrikes()) {
             System.out.println(endMessage);
             String restart = Console.readLine();
